Added new state for working on Lua script testing. Complete rework of serialization to be able handle serializing components to a file or to Lua. Switched images from using an int as an identifier to an uint.

+++ src/Game/Graphics/triRendering.c
@@ -154,7 +154,7 @@ static bool testTriangle( Vector2* p0, Vector2* p1, Vector2* p2 )
 }
 
 static int addTriangle( TriangleList* triList, TriVert vert0, TriVert vert1, TriVert vert2,
-	ShaderType shader, PlatformTexture texture, PlatformTexture extraTexture, float floatVal0, int clippingID, uint32_t camFlags, int8_t depth )
+	ShaderType shader, PlatformTexture texture, PlatformTexture extraTexture, float floatVal0, int8_t clippingID, uint32_t camFlags, int8_t depth )
 {
 	// test to see if the triangle can be culled
 	bool anyInside = false;
@@ -212,13 +212,13 @@ We'll assume the array has three vertices in it.
  Return a value < 0 if there's a problem.
 */
 int triRenderer_AddVertices( TriVert* verts, ShaderType shader, PlatformTexture texture, PlatformTexture extraTexture,
-	float floatVal0, int clippingID, uint32_t camFlags, int8_t depth, TriType type )
+	float floatVal0, int8_t clippingID, uint32_t camFlags, int8_t depth, TriType type )
 {
 	return triRenderer_Add( verts[0], verts[1], verts[2], shader, texture, extraTexture, floatVal0, clippingID, camFlags, depth, type );
 }
 
 int triRenderer_Add( TriVert vert0, TriVert vert1, TriVert vert2, ShaderType shader, PlatformTexture texture, PlatformTexture extraTexture,
-	float floatVal0, int clippingID, uint32_t camFlags, int8_t depth, TriType type )
+	float floatVal0, int8_t clippingID, uint32_t camFlags, int8_t depth, TriType type )
 {
 	switch( type ) {
 	case TT_SOLID:
